feat(preset-umi): external dependencies in ssr builder (#10213)
* feat(preset-umi): external dependencies in ssr builder

* refactor: use `isMonorepo`

* refactor: move `isMonorepo` to utils package

* style: merge import

* Update packages/preset-umi/src/features/ssr/builder/builder.ts

* feat(preset-umi): external devDependencies in ssr builder
