+++ packages/preset-umi/src/features/monorepo/redirect.ts
@@ -1,4 +1,4 @@
-import { isLocalDev, logger } from '@umijs/utils';
+import { isLocalDev, logger, isMonorepo } from '@umijs/utils';
 import { pkgUp } from '@umijs/utils/compiled/pkg-up';
 import assert from 'assert';
 import { existsSync, statSync } from 'fs';
@@ -110,23 +110,5 @@ async function collectAllProjects(opts: IOpts) {
       return obj;
     },
     {},
-  );
-}
-
-const MONOREPO_FILE = ['pnpm-workspace.yaml', 'lerna.json'];
-export function isMonorepo(opts: IOpts) {
-  const pkgPath = join(opts.root, 'package.json');
-  let pkg: Record<string, any> = {};
-  try {
-    pkg = require(pkgPath);
-  } catch (e) {}
-  const pkgExist = existsSync(pkgPath);
-  return (
-    pkgExist &&
-    (MONOREPO_FILE.some((file) => {
-      return existsSync(join(opts.root, file));
-    }) ||
-      // npm workspaces
-      pkg?.workspaces)
   );
 }
