+++ packages/preset-umi/src/features/ssr/builder/builder.ts
@@ -1,5 +1,5 @@
 import esbuild from '@umijs/bundler-utils/compiled/esbuild';
-import { logger } from '@umijs/utils';
+import { isMonorepo, logger } from '@umijs/utils';
 import { resolve } from 'path';
 import { IApi } from '../../../types';
 import { absServerBuildPath, esbuildUmiPlugin } from '../utils';
@@ -13,8 +13,18 @@ export async function build(opts: { api: IApi; watch?: boolean }) {
   logger.wait('[SSR] Compiling...');
   const now = new Date().getTime();
 
+  function getExternal() {
+    // TODO: 先不考虑 monorepo 的场景，因为可能依赖了其他的子包
+    if (isMonorepo({ root: api.paths.cwd })) {
+      return [];
+    } else {
+      const dependencies = Object.keys(api.pkg.dependencies || {});
+      const devDependencies = Object.keys(api.pkg.devDependencies || {});
+      return [...dependencies, ...devDependencies];
+    }
+  }
+
   // TODO: 支持通用的 alias
-  // TODO: external all import from package.json.dependencies
   await esbuild.build({
     format: 'cjs',
     platform: 'node',
@@ -43,6 +53,7 @@ export async function build(opts: { api: IApi; watch?: boolean }) {
       assetsLoader({ cwd: api.cwd }),
     ],
     outfile: absServerBuildPath(api),
+    external: getExternal(),
   });
   const diff = new Date().getTime() - now;
   logger.info(`[SSR] Compiled in ${diff}ms`);
