+++ packages/utils/src/index.ts
@@ -42,6 +42,7 @@ export * as register from './register';
 export * from './tryPaths';
 export * from './winPath';
 export * from './setNoDeprecation';
+export * from './isMonorepo';
 export {
   address,
   axios,
